Improve ndarray testing
